Added adaptave stepsizes and more example metrics
Added more example metrics and changed the t to s in the GeneralGRcode.
Tested out new metrics a little bit.

Added adaptive step sizes to the inital value problem solver in
numbericalMethods.py. Did not thouroughly test but it seems to work very
well on the examples I did test.