Added unit tests
Previously tests were just commented out when not in use. Moved some of
these to unit test files. There is still more work required to finish
this.

Also reworked metrics so that they can all be uncommented at the same
time. Now there is  just a function named metric which can be set to any
of them or defined as a new metric in some other function.

Lastly moved the tolerance to the inputs of initialValueSoution in
numericalMethods.