Added function which aids in finding null geodesics
Might rename function later, but the function currently called nullTest
should return zero if the coordinates it is evaluated correspond to a
point on a null geodesic.

Null geodesics are difficult to maintain exactly since The initial point
can't be any closer than machine epsilon to the null geodesic and plenty
of error is introduced while numerical integration is happening.