Got rid of global constants
Was previously using global constants to set constants in different
metrics (for example rs in the Schwarzchild metric). Changed this in
favor of having the constants be input into a function which returns the
metric function. All the tests and examples had to be changed since they
all used the old method of choosing a metric. Many small bugs were fixed
as well, the most interesting of which was that the Kerr metric was
accidnetally using the global r constant instead of its input rad
variable in one small spot. This interstingly didn't have as big of a
change as I would have expected for the KerrMetricTest.py example.